requirements.txt
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.1.70
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
apscheduler>=3.8.0
ccxt>=1.80.0
gunicorn>=20.1.0
python-dotenv>=0.19.0